refactor: refactor code structure according to feedbacks
