+++ library/lib/edges/index.ts
@@ -1,2 +1,9 @@
-export * from "./edges" // Custom edge components
-export * from "./types" // Custom edge types
+
+export * from "./types"
+export * from "./AggregationEdge"
+export * from "./BiAssociationEdge"
+export * from "./CompositionEdge"
+export * from "./DependencyEdge"
+export * from "./InheritanceEdge"
+export * from "./RealizationEdge"
+export * from "./UniAssociationEdge"
