+++ library/lib/edges/types.tsx
@@ -1,13 +1,11 @@
 import { EdgeTypes } from "@xyflow/react"
-import {
-  InheritanceEdge,
-  RealizationEdge,
-  AggregationEdge,
-  CompositionEdge,
-  BiAssociationEdge,
-  UniAssociationEdge,
-  DependencyEdge,
-} from "./edges"
+import { InheritanceEdge } from "./InheritanceEdge"
+import { RealizationEdge } from "./RealizationEdge"
+import { AggregationEdge } from "./AggregationEdge"
+import { CompositionEdge } from "./CompositionEdge"
+import { BiAssociationEdge } from "./BiAssociationEdge"
+import { UniAssociationEdge } from "./UniAssociationEdge"
+import { DependencyEdge } from "./DependencyEdge"
 export const diagramEdgeTypes = {
   inheritance: InheritanceEdge,
   realization: RealizationEdge,
