+++ library/lib/hooks/useReconnect.ts
@@ -1,0 +1,26 @@
+import {
+    reconnectEdge,
+    Edge,
+    Connection,
+    useReactFlow
+  } from "@xyflow/react"
+import { useCallback} from "react"
+
+export const useReconnect = () => {
+
+  const { setEdges} = useReactFlow()
+  const onReconnect = useCallback(
+    (oldEdge: Edge, newConnection: Connection) =>
+      setEdges((els) =>
+        reconnectEdge(oldEdge, newConnection, els).map((edge) => ({
+          ...edge,
+          // Preserve existing marker or assign a default
+          type: edge.type,
+          markerEnd: edge.markerEnd,
+        }))
+      ),
+    []
+  )
+  return { onReconnect }
+}
+
